Add files via upload
This is a one-page portfolio website showcasing some of my skills, projects, and professional experience. It shows my personal brand, my abilities, and aspirations. I also learned a lot when preparing this portfolio, from using different fonts and text decorations to the text weights, and colors, and handling the images.